+++ src/main/java/ru/j3v/io/DataReader.java
@@ -1,9 +1,10 @@
 package ru.j3v.io;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.TreeMap;
 
 public interface DataReader {
-    TreeMap<Date, Double> readStocks();
-    TreeMap<Date, Double> readInfl();
+    TreeMap<Date, BigDecimal> readStocks();
+    TreeMap<Date, BigDecimal> readInfl();
 }
